+++ library/Zend/Mail/Header/Sender.php
@@ -53,6 +53,8 @@ class Sender implements HeaderInterface
                 $senderName = null;
             }
             $senderEmail = $matches['email'];
+        } else {
+            $senderEmail = $value;
         }
 
         $header->setAddress($senderEmail, $senderName);